Moved to cmake

+++ src/main.cpp
@@ -1,21 +1,24 @@
 #include <iostream>
 
-#include "Quicksort.h"
-#include "BubbleSort.h"
-#include "SortVisualiser.h"
+#include "SortingAlgorithms/Quicksort.h"
+#include "SortingAlgorithms/BubbleSort.h"
+#include "SortVisualiser/SortVisualiser.h"
 
 int main()
 {
     const int SAMPLES = 1000;
     const int WIDTH = 1000;
     const int HEIGHT = 800;
-    const int DELAY = 2;
+    const float DELAY = 2;
     std::string name = "Quicksort";
     
-    // Press S to start
-    // It is built so you can change algorithms very easy.
-    SortVisualiser<Quicksort> SortVis = SortVisualiser<Quicksort>(WIDTH, HEIGHT, SAMPLES, DELAY, name);
-
+    SortVisualiser<Quicksort> SortVis = SortVisualiser<Quicksort>(
+        WIDTH, 
+        HEIGHT, 
+        SAMPLES, 
+        DELAY, 
+        name
+    );
 
     return 0;
 }
